Uploaded Modified Project 2
In a new folder, uploaded the modified project 2 code that uses transformer encoder decoder on the old dataset. The loss curves and generation are not good because the data is not built for the model, but they do work, so I am hoping the new data that is more suited for this model will make it good.